Makes the compilation process less noisy.

+++ lib/qdp_mastermap.cc
@@ -138,28 +138,28 @@ namespace QDP {
       }
     return id;
   }
-
+  
   const multi1d<int>& MasterMap::getInnerSites(const Subset& s,int bitmask) const {
-    assert( s.getId() >= 0 && s.getId() < powerSetC.size()               && "subset Id out of range");
-    assert( bitmask   >  0 && bitmask   < powerSetC.at(s.getId()).size() && "bitmask out of range");
+    assert( s.getId() >= 0 && s.getId() < (int)powerSetC.size()               && "subset Id out of range");
+    assert( bitmask   >  0 && bitmask   < (int)powerSetC.at(s.getId()).size() && "bitmask out of range");
     return *powerSetC[s.getId()][bitmask];
   }
 
   const multi1d<int>& MasterMap::getFaceSites(const Subset& s,int bitmask) const {
-    assert( s.getId() >= 0 && s.getId() < powerSet.size()               && "subset Id out of range");
-    assert( bitmask   >  0 && bitmask   < powerSet.at(s.getId()).size() && "bitmask out of range");
+    assert( s.getId() >= 0 && s.getId() < (int)powerSet.size()               && "subset Id out of range");
+    assert( bitmask   >  0 && bitmask   < (int)powerSet.at(s.getId()).size() && "bitmask out of range");
     return *powerSet[s.getId()][bitmask]; 
   }
 
   int MasterMap::getCountInner(const Subset& s,int bitmask) const {
-    assert( s.getId() >= 0 && s.getId() < powerSetC.size()               && "subset Id out of range");
-    assert( bitmask   >  0 && bitmask   < powerSetC.at(s.getId()).size() && "bitmask out of range");
+    assert( s.getId() >= 0 && s.getId() < (int)powerSetC.size()               && "subset Id out of range");
+    assert( bitmask   >  0 && bitmask   < (int)powerSetC.at(s.getId()).size() && "bitmask out of range");
     return powerSetC[s.getId()][bitmask]->size();
   }
 
   int MasterMap::getCountFace(const Subset& s,int bitmask) const {
-    assert( s.getId() >= 0 && s.getId() < powerSet.size()                && "subset Id out of range");
-    assert( bitmask   >  0 && bitmask   < powerSet.at(s.getId()).size()  && "bitmask out of range");
+    assert( s.getId() >= 0 && s.getId() < (int)powerSet.size()                && "subset Id out of range");
+    assert( bitmask   >  0 && bitmask   < (int)powerSet.at(s.getId()).size()  && "bitmask out of range");
     return powerSet[s.getId()][bitmask]->size(); 
   }
